pyproject.toml
[project]
name = "molmoweb"
version = "0.1.0"
description = "MolmoWeb - Web agent evaluation framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=24.10.0",
    "browserbase==1.4.0",
    "fastapi>=0.115.12",
    "fire>=0.7.0",
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
    "isort>=5.13.2",
    "jinja2>=3.1.5",
    "openai>=2.8.1",
    "playwright==1.44",
    "python-fasthtml>=0.12.4",
    "tenacity==9.1.2",
    "transformers>=4.51,<5",
    "ai2-molmo2 @ git+https://github.com/allenai/molmo2.git",
    "google-genai>=1.49.0",
    "molmo-utils>=0.0.1",
    "torchmetrics>=1.0.0",
    "backoff>=2.0.0",
    "marimo>=0.21.1",
    "pyright>=1.1.408",
]

[tool.isort]
profile = "black"
